+++ InnocentGrey/IGKit/igatool/igatool.vcxproj
@@ -10,6 +10,9 @@
       <Platform>Win32</Platform>
     </ProjectConfiguration>
   </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="igatool.cpp" />
+  </ItemGroup>
   <PropertyGroup Label="Globals">
     <ProjectGuid>{C45FDDBD-8A1E-4726-AA47-E4C6188649DA}</ProjectGuid>
     <RootNamespace>igatool</RootNamespace>
@@ -66,12 +69,6 @@
       <OptimizeReferences>true</OptimizeReferences>
     </Link>
   </ItemDefinitionGroup>
-  <ItemGroup>
-    <ClCompile Include="exiga.cpp" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="as-util.h" />
-  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
